refactor(module-product): Product 도메인 모듈로 변경
