+++ module-product/src/main/java/kitchenpos/product/application/ProductService.java
@@ -3,9 +3,9 @@ package kitchenpos.product.application;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.product.domain.Product;
+import kitchenpos.product.repository.ProductRepository;
 import kitchenpos.product.dto.ProductRequest;
 import kitchenpos.product.dto.ProductResponse;
-import kitchenpos.product.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
