+++ module-product/src/main/java/kitchenpos/product/controller/ProductController.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.ui;
+package kitchenpos.product.controller;
 
 import java.net.URI;
 import java.util.List;
