+++ module-product/src/main/java/kitchenpos/product/dto/ProductRequest.java
@@ -1,11 +1,12 @@
 package kitchenpos.product.dto;
 
-import static kitchenpos.common.ValidationMessage.NOT_EMPTY;
-import static kitchenpos.common.ValidationMessage.POSITIVE_OR_ZERO;
+
+import static kitchenpos.common.message.ValidationMessage.NOT_EMPTY;
+import static kitchenpos.common.message.ValidationMessage.POSITIVE_OR_ZERO;
 
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.PositiveOrZero;
-import kitchenpos.menu.domain.Price;
+import kitchenpos.common.domain.Price;
 import kitchenpos.product.domain.Product;
 
 public class ProductRequest {
