+++ module-product/src/test/java/kitchenpos/product/ModuleProductApplicationTests.java
@@ -1,0 +1,13 @@
+package kitchenpos.product;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ModuleProductApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
