+++ module-product/src/test/java/kitchenpos/product/domain/ProductTest.java
@@ -2,7 +2,7 @@ package kitchenpos.product.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import kitchenpos.menu.domain.Price;
+import kitchenpos.common.domain.Price;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
